[IOPLT-599] Replace legacy StatusContent with Alert component for SectionStatus advices
